LayoutTests/media/csp-blocks-video-expected.txt

CONSOLE MESSAGE: Refused to load media from 'test.mp4' because it violates the following Content Security Policy directive: "media-src 'none'".

END OF TEST
This test passes if it doesn't alert failure.
